adds dummy extractor
The extractor simply generates a message saying that sitebag was
unable to extract text from the given document. This is useful if the
original document should nevertheless be stored in sitebag, even
without content extraction. As more extractors may come in over time,
extraction could be retried later on this document.